feat: Rewrite extension to plain JavaScript
This commit rewrites the entire extension from TypeScript to plain JavaScript, removing the need for a build system. This simplifies the project structure and makes it easier to get started with the code.

Changes include:
- All TypeScript files have been converted to JavaScript.
- The module system has been removed, and files are loaded in order via the `manifest.json`.
- The `src`, `dist`, and `old` directories have been replaced by a single `extension` directory that contains the unpacked extension.
- Dependencies and build scripts have been removed.
- A new `README.md` has been added to the `extension` directory with updated instructions.
