+++ extension/manifest.json
@@ -1,0 +1,24 @@
+{
+  "manifest_version": 3,
+  "name": "UI Viz Component Highlighter",
+  "version": "2.0",
+  "description": "Highlights UI components and adds links based on JSON mapping.",
+  "permissions": ["activeTab", "scripting", "storage"],
+  "background": {
+    "service_worker": "js/background.js"
+  },
+  "content_scripts": [
+    {
+      "matches": ["*://*/*"],
+      "js": ["js/defaultSettings.js", "js/contentScript.js"],
+      "run_at": "document_end"
+    }
+  ],
+  "action": {
+    "default_title": "Toggle Component Highlights",
+    "default_icon": "icon.png"
+  },
+  "icons": {
+    "128": "icon.png"
+  }
+}
