--- old/manifest.json
@@ -1,27 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "UI Viz Component Highlighter",
-  "version": "1.0",
-  "description": "Highlights UI components and adds links based on JSON mapping.",
-  "permissions": ["activeTab", "scripting"],
-  "background": {
-    "service_worker": "background.js"
-  },
-  "content_scripts": [
-    {
-      "matches": ["*://*.siemens.com/*", "file:///*"],
-      "js": ["contentScript.js"],
-      "run_at": "document_end"
-    }
-  ],
-  "web_accessible_resources": [
-    {
-      "resources": ["injectedScript.js", "styles.css"],
-      "matches": ["*://*.siemens.com/*", "file:///*"]
-    }
-  ],
-  "action": {
-    "default_title": "Toggle Component Highlights",
-    "default_icon": "icon.png"
-  }
-}
