--- old/popup.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>UI Component Highlighter</title>
-  <style>
-    body { width: 200px; }
-  </style>
-</head>
-<body>
-  <h1>UI Component Highlighter</h1>
-  <p>Components highlighted on this page.</p>
-</body>
-</html>
