--- old/section-mapper.json
@@ -1,59 +1,0 @@
-{
-  "components": [
-    {
-      "name": "Banner",
-      "contentTypeUrl": "https://docs.example.com/FooterComponent",
-      "uxDocsUrl": "",
-      "identifiers": {
-        "className": "sitesBanner",
-        "id": "730VWGE2yHuzuKqADigbb3"
-      },
-      "borderHighlightStyle": "1px dotted red"
-    },
-    {
-      "name": "Hero",
-      "contentTypeUrl": "https://docs.example.com/HeroComponent",
-      "uxDocsUrl": "",
-      "identifiers": {
-        "className": "heroComponent"
-      },
-      "borderHighlightStyle": "1px solid red"
-    },
-    {
-      "name": "Cards Container",
-      "contentTypeUrl": "https://docs.example.com/HeroComponent",
-      "uxDocsUrl": "",
-      "identifiers": {
-        "className": "sitesCardsContainer"
-      },
-      "borderHighlightStyle": "1px dotted green"
-    },
-    {
-      "name": "Cards",
-      "contentTypeUrl": "https://docs.example.com/HeroComponent",
-      "uxDocsUrl": "",
-      "identifiers": {
-        "className": "siemens-cards"
-      },
-      "borderHighlightStyle": "1px dotted blue"
-     },
-     {
-      "name": "3 Col Tiles",
-      "contentTypeUrl": "https://docs.example.com/HeroComponent",
-      "uxDocsUrl": "",
-      "identifiers": {
-        "className": "sitesThreeColTiles"
-      },
-      "borderHighlightStyle": "1px dotted blue"
-    },
-    {
-      "name": "Promo",
-      "contentTypeUrl": "https://docs.example.com/HeroComponent",
-      "uxDocsUrl": "",
-      "identifiers": {
-        "className": "promo-section"
-      },
-      "borderHighlightStyle": "1px dotted blue"
-    } 
-  ]
-}
