--- src/styles/styles.css
@@ -1,5 +1,0 @@
-/* styles.css */
-
-.extension-component-highlight-reset-style { 
-  all: revert;
-}
